requirements.txt
pymongo==4.8.0
tqdm==4.66.4
dnspython==2.6.1
neo4j==5.23.0
python-dotenv==1.0.1
